refactor(Runtime): extend `RenderView` and implement `DynamicView`
Preivously, `AppView` extended `DynamicView` and implemented `RenderView`, but
it's become clear that there's greater value in extending `RenderView`.

PiperOrigin-RevId: 245806262
